Added tasks for lessons 11, 12, 13
