+++ src/lesson_11/AverageLineLength_3.java
@@ -16,12 +16,14 @@ public class AverageLineLength_3 {
         String str3 = sc.nextLine();
 
         String[] strs = {str1, str2, str3};
+
         int totalLength = str1.length() + str2.length() + str3.length();
         double averageLength = totalLength / 3.0;
 
         System.out.println("Средняя длина: " + averageLength);
 
         System.out.println("Строки, длина которых меньше средней:");
+
         for (String str : strs) {
             if (str.length() < averageLength) {
                 System.out.println(str + " (длина: " + str.length() + ")");
